Add files via upload
The objective of this analysis is to develop a machine learning model aimed at predicting injury risk in athletes. Once the model is constructed, a thorough analysis will be conducted to identify the most influential variables in determining injury risk. Additionally, the analysis will explore the optimal thresholds at which these variables should be adjusted to mitigate the likelihood of injury.

The model was developed using a dataset sourced from Kaggle, which contains information on collegiate athlete injury rates and performance metrics.